backend/requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
gunicorn
Flask-Migrate
pytest
pytest-cov
stripe
APScheduler
sendgrid